LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/focus-prioritized-expected.txt


FAIL Anchor selection prioritized focused element. assert_equals: y coordinate expected 458 but got 3358
